+++ lib/langtool/accounts/user.ex
@@ -21,11 +21,19 @@ defmodule Langtool.Accounts.User do
     |> unique_constraint(:email)
     |> update_change(:encrypted_password, &Bcrypt.hashpwsalt/1)
     |> update_change(:confirmed_at, nil)
+    |> update_change(:confirmation_token, random_string(24))
   end
 
   @doc false
   def changeset(user, attrs) do
     user
     |> cast(attrs, [:confirmed_at])
+  end
+
+  defp random_string(length) do
+    length
+    |> :crypto.strong_rand_bytes()
+    |> Base.url_encode64()
+    |> binary_part(0, length)
   end
 end