+++ lib/langtool_web/controllers/preview_email_controller.ex
@@ -1,13 +1,20 @@
 defmodule LangtoolWeb.PreviewEmailController do
   use LangtoolWeb, :controller
-  alias Langtool.{Email}
+  alias Langtool.{Accounts.User}
+  alias LangtoolWeb.{UserMailer}
 
   def show(conn, %{"name" => name, "type" => type}) do
-    email = Email.render_email(name, type, "us@example.com")
-    render_email(conn, email, type)
+    name
+    |> build_email()
+    |> render_email(conn, type)
   end
 
-  defp render_email(conn, email, type) do
+  defp build_email("welcome_email") do
+    %User{email: "john.doe@example.com", confirmation_token: "1234567890"}
+    |> UserMailer.welcome_email()
+  end
+
+  defp render_email(email, conn, type) do
     conn
     |> Plug.Conn.put_resp_content_type("text/html")
     |> send_resp(:ok, email_body(email, type))